feat: add delete customer controller

+++ src/controllers/delete-customer.controller.ts
@@ -2,12 +2,16 @@ import { Request, Response } from 'express';
 import { CustomerModel } from '../models/customer.model';
 import { StatusCodes } from 'http-status-codes';
 import errorHandler from '../utils/error-handler.util';
+import BadRequestError from '../errors/bad-request.error';
 import NotFoundError from '../errors/not-found.error';
 
 export const deleteCustomer = async (req: Request, res: Response) => {
     try {
         const customerId = req.params.id;
 
+        if (!customerId) {
+            throw new BadRequestError('Invalid customer ID');
+        }
         const deletedCustomer =
             await CustomerModel.findByIdAndDelete(customerId);
